more minor C++ fixes

+++ cudax/test/stf/loop_dispatch/loop_dispatch.cu
@@ -29,7 +29,7 @@ int main()
     auto lA = ctx.logical_data<int>(size_t(1024 * 1024));
 
     ctx.parallel_for(ctx.current_exec_place(), lA.shape(), lA.write())->*[] __device__(size_t i, auto a) {
-      a(i) = (int) (10.0 * cos(i));
+      a(i) = (int) (10.0 * cos((double)i));
     };
 
     ctx.parallel_for(ctx.current_exec_place(), lA.shape(), lA.rw(), lB.read())->*[] __device__(size_t i, auto a, auto b) {
